+++ controllers/utilities/error.js
@@ -8,7 +8,7 @@ class ExtendableError extends Error {
     if (Reflect.getPrototypeOf(new.target) === TransferableError) {                                                     // note: lock-in this parental class can be only extended by a special child class
       super();
       this.name = this.constructor.name;
-      this.message = errorCodeProxyAgent[this.name](entry, literals);
+      this.message = String(errorCodeProxyAgent[this.name](entry, literals));
       if (checkToStringTag(entry, 'Number')) this.code = entry;
       if (Error.captureStackTrace) Error.captureStackTrace(this, this.constructor);                                     // note: V8 JS-engine only
       else this.stack = (new Error(this.message)).stack;                                                                // note: non-V8 browser only